Make parts inventory quantity/reorder fields integers
quantity_in_stock, reorder_level, and reorder_unit are counted in whole
units (you don't receive 0.5 pads), so DECIMAL(10,2) didn't fit. reorder_cost
and markup are unaffected -- those are genuinely fractional (money, a
multiplier).

- 01-schema.sql: these three columns are now INT (fresh deploys only).
- New docker/init-db/migrate-parts-integer-columns.sql: one-off ALTER
  for already-deployed databases, following the same pattern as the
  earlier AUTO_INCREMENT migration. Applied and verified against the
  local test database.
- backend/server.js: POST /api/parts, PUT /api/parts/:id, and
  POST /api/parts/:id/receive now round these fields via a toIntOrNull
  helper instead of trusting the client to send whole numbers.
- frontend: the corresponding number inputs use step="1" instead of
  step="0.01" in both the Receive Parts form and the part edit page.
- Verified end-to-end: fractional values sent to the API (133.7, 7.9,
  2.2, 10.4) all round correctly and persist as integers; reorder_cost
  stays decimal.
- PROJECT_STATUS.md: documented the GET/PUT /api/parts/:id routes that
  were missing from the previous commit, plus this change.

+++ docker/init-db/01-schema.sql
@@ -77,10 +77,10 @@ CREATE TABLE parts_inventory (
   part_name VARCHAR(255) NOT NULL,
   description TEXT,
   category VARCHAR(100),                  -- Pads, Supplies, Cork, Mouthpiece Patch, ...
-  quantity_in_stock DECIMAL(10,2),
-  reorder_level DECIMAL(10,2),
+  quantity_in_stock INT,
+  reorder_level INT,
   reorder_cost DECIMAL(10,2),
-  reorder_unit DECIMAL(10,2),
+  reorder_unit INT,
   reorder_url VARCHAR(500),
   markup DECIMAL(6,2),
   vendor_id INT,
